requirements.txt
django
djangorestframework
markdown
django-filter
Pillow
djangorestframework-simplejwt
razorpay
stripe
django-cors-headers
python-decouple
django-ckeditor
